Add session handling to config module
This commit introduces a new class, Session, to the config module. This class is responsible for reading and writing session data to a file. The session file location can be configured via an environment variable. The session data is now also included in the output of the read_config function. A test case for the session read and write functionality has been added.